--- ECDL-Escape Game/Assets/Scripts/MenuScript/PauseManu.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 6cce8e521d08a40499e9d5da8f0b9950
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
